fix: 修复decode确定性、添加贪心初始化、instance_parser支持浮点、main支持批量测试
1. ga.py:
   - decode使用模运算代替random.randint确保确定性
   - 新增_greedy_ms贪心机器选择初始化(30%纯贪心+30%纯混沌+40%混合)
   - 调整变异率: ms_pm = min(pm*1.5, 0.5)
   - 添加gen_progress参数用于未来自适应变异

2. alns.py:
   - decode同样使用模运算确保确定性

3. config.py:
   - DATA_ROOT调整为直接指向Brandimarte_Data目录

4. core/instance_parser.py:
   - 支持浮点加工时间(如3.5)

5. main.py:
   - 支持python main.py all批量跑Mk01-Mk10
   - 支持python main.py Mk02简写
   - 输出汇总表格

+++ core/instance_parser.py
@@ -40,9 +40,9 @@ def load_fjsp_from_file(filepath):
     first_line = lines[idx].strip()
     idx += 1
     
-    parts = list(map(int, first_line.split()))
-    num_jobs = parts[0]
-    num_machines = parts[1]
+    parts = list(map(float, first_line.split()))
+    num_jobs = int(parts[0])
+    num_machines = int(parts[1])
     # 第三个数字（平均可选机器数）可忽略
     
     jobs = []
@@ -57,9 +57,12 @@ def load_fjsp_from_file(filepath):
         data_line = lines[idx].strip()
         idx += 1
         
-        data = list(map(int, data_line.split()))
+
+
+
+        data = list(map(float, data_line.split()))
         
-        num_ops = data[0]
+        num_ops = int(data[0])
         job_operations = []
         pos = 1
         
@@ -70,7 +73,8 @@ def load_fjsp_from_file(filepath):
                 idx += 1
                 data.extend(map(int, next_line.split()))
             
-            k = data[pos]
+
+            k = int(data[pos])
             pos += 1
             
             if pos + 2*k > len(data):
@@ -80,9 +84,12 @@ def load_fjsp_from_file(filepath):
             
             machines = []
             times = []
+
+
+
             for _ in range(k):
-                machine_id = data[pos]
-                proc_time = data[pos+1]
+                machine_id = int(data[pos])
+                proc_time = float(data[pos+1])
                 pos += 2
                 machines.append(machine_id)
                 times.append(proc_time)
